+++ backend/routes/api/session.js
@@ -12,42 +12,42 @@ const router = express.Router();
 
 // Log in
 // POST -> /api/session/
-router.post('/', validateLogin, async (req, res, next) => {
-  const { credential, password } = req.body;
+// router.post('/', validateLogin, async (req, res, next) => {
+//   const { credential, password } = req.body;
 
-  const user = await User.unscoped().findOne({
-      // allows either username or password to log in
-      where: {
-        [Op.or]: {
-          username: credential,
-          email: credential
-        }
-      }
-    });
+//   const user = await User.unscoped().findOne({
+//       // allows either username or password to log in
+//       where: {
+//         [Op.or]: {
+//           username: credential,
+//           email: credential
+//         }
+//       }
+//     });
 
-    if (!user || !bcrypt.compareSync(password, user.hashedPassword.toString())) {
-      const err = new Error('Login failed');
-      err.status = 401;
-      err.title = 'Login failed';
-      err.errors = { credential: 'The provided credentials were invalid.' };
-      return next(err);
-    }
+//     if (!user || !bcrypt.compareSync(password, user.hashedPassword.toString())) {
+//       const err = new Error('Login failed');
+//       err.status = 401;
+//       err.title = 'Login failed';
+//       err.errors = { credential: 'The provided credentials were invalid.' };
+//       return next(err);
+//     }
 
-    const safeUser = {
-      id: this.id,
-      username: this.username,
-      email: this.email,
-      firstName: this.firstName,
-      lastName: this.lastName,
-    };
+//     const safeUser = {
+//       id: this.id,
+//       username: this.username,
+//       email: this.email,
+//       firstName: this.firstName,
+//       lastName: this.lastName,
+//     };
 
-    await setTokenCookie(res, safeUser);
+//     await setTokenCookie(res, safeUser);
 
-    return res.json({
-      user: safeUser
-    });
-  }
-);
+//     return res.json({
+//       user: safeUser
+//     });
+//   }
+// );
 
 // Log out
 router.delete('/', (_req, res) => {
