+++ backend/utils/validators/users.js
@@ -29,6 +29,12 @@ const validateSignup = [
     .exists({ checkFalsy: true })
     .isLength({ min: 6 })
     .withMessage('Password must be 6 characters or more.'),
+  check('firstName')
+  .exists({ checkFalsy: true })
+  .withMessage('Please provide a first name'),
+  check('lastName')
+  .exists({ checkFalsy: true })
+  .withMessage('Please provide a last name'),
   handleValidationErrors
 ];
